requirements.txt
fastapi
uvicorn
pdfplumber
fpdf
sentence-transformers
google-generativeai
Flask
scikit-learn
PyMuPDF
python-dotenv
weasyprint